// src/types/Ancrages.ts
interface Ancrages {
  hats?: [x: number, y: number, z: number];
  heads?: [x: number, y: number, z: number];
  costumes?: [x: number, y: number, z: number];
  hand_l?: [x: number, y: number, z: number];
  hand_r?: [x: number, y: number, z: number];
  foot_l?: [x: number, y: number, z: number];
  foot_r?: [x: number, y: number, z: number];
}

export type { Ancrages };